+++ core/ivr/ivr.go
@@ -27,6 +27,7 @@ import (
 	"github.com/nyaruka/mailroom/core/models"
 	"github.com/nyaruka/mailroom/core/runner"
 	"github.com/nyaruka/mailroom/runtime"
+	"github.com/nyaruka/mailroom/utils/clogs"
 	"github.com/nyaruka/null/v3"
 )
 
@@ -128,7 +129,7 @@ func HangupCall(ctx context.Context, rt *runtime.Runtime, call *models.Call) (*m
 		clog.HTTP(trace)
 	}
 	if err != nil {
-		clog.Error(err)
+		clog.Error(clogs.NewLogError("", "", err.Error()))
 	}
 
 	if err := call.AttachLog(ctx, rt.DB, clog); err != nil {
@@ -260,7 +261,7 @@ func RequestStartForCall(ctx context.Context, rt *runtime.Runtime, channel *mode
 		clog.HTTP(trace)
 	}
 	if err != nil {
-		clog.Error(err)
+		clog.Error(clogs.NewLogError("", "", err.Error()))
 
 		// set our status as errored
 		err := call.UpdateStatus(ctx, rt.DB, models.CallStatusFailed, 0, time.Now())
@@ -302,7 +303,7 @@ func StartIVRFlow(
 	}
 
 	// get the flow for our start
-	start, err := models.GetFlowStartAttributes(ctx, rt.DB, startID)
+	start, err := models.GetFlowStartByID(ctx, rt.DB, startID)
 	if err != nil {
 		return fmt.Errorf("unable to load start: %d: %w", startID, err)
 	}
@@ -607,7 +608,7 @@ func HandleIVRStatus(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAss
 		}
 
 		// on errors we need to look up the flow to know how long to wait before retrying
-		start, err := models.GetFlowStartAttributes(ctx, rt.DB, call.StartID())
+		start, err := models.GetFlowStartByID(ctx, rt.DB, call.StartID())
 		if err != nil {
 			return fmt.Errorf("unable to load start: %d: %w", call.StartID(), err)
 		}