+++ core/runner/runner.go
@@ -5,6 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"log/slog"
+	"maps"
+	"slices"
 	"time"
 
 	"github.com/gomodule/redigo/redis"
@@ -16,7 +18,6 @@ import (
 	"github.com/nyaruka/mailroom/core/goflow"
 	"github.com/nyaruka/mailroom/core/models"
 	"github.com/nyaruka/mailroom/runtime"
-	"golang.org/x/exp/maps"
 )
 
 const (
@@ -122,49 +123,37 @@ func ResumeFlow(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, 
 }
 
 // StartFlowBatch starts the flow for the passed in org, contacts and flow
-func StartFlowBatch(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, batch *models.FlowStartBatch) ([]*models.Session, error) {
-	start := time.Now()
-
-	// if this is our last start, no matter what try to set the start as complete as a last step
-	if batch.IsLast {
-		defer func() {
-			err := models.MarkStartComplete(ctx, rt.DB, batch.StartID)
-			if err != nil {
-				slog.Error("error marking start as complete", "error", err, "start_id", batch.StartID)
-			}
-		}()
-	}
-
+func StartFlowBatch(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, start *models.FlowStart, batch *models.FlowStartBatch) ([]*models.Session, error) {
 	// try to load our flow
-	flow, err := oa.FlowByID(batch.FlowID)
+	flow, err := oa.FlowByID(start.FlowID)
 	if err == models.ErrNotFound {
-		slog.Info("skipping flow start, flow no longer active or archived", "flow_id", batch.FlowID)
+		slog.Info("skipping flow start, flow no longer active or archived", "flow_id", start.FlowID)
 		return nil, nil
 	}
 	if err != nil {
-		return nil, fmt.Errorf("error loading campaign flow: %d: %w", batch.FlowID, err)
+		return nil, fmt.Errorf("error loading flow: %d: %w", start.FlowID, err)
 	}
 
 	// get the user that created this flow start if there was one
 	var flowUser *flows.User
-	if batch.CreatedByID != models.NilUserID {
-		user := oa.UserByID(batch.CreatedByID)
+	if start.CreatedByID != models.NilUserID {
+		user := oa.UserByID(start.CreatedByID)
 		if user != nil {
 			flowUser = oa.SessionAssets().Users().Get(user.Email())
 		}
 	}
 
 	var params *types.XObject
-	if !batch.Params.IsNull() {
-		params, err = types.ReadXObject(batch.Params)
+	if !start.Params.IsNull() {
+		params, err = types.ReadXObject(start.Params)
 		if err != nil {
 			return nil, fmt.Errorf("unable to read JSON from flow start params: %w", err)
 		}
 	}
 
 	var history *flows.SessionHistory
-	if !batch.SessionHistory.IsNull() {
-		history, err = models.ReadSessionHistory(batch.SessionHistory)
+	if !start.SessionHistory.IsNull() {
+		history, err = models.ReadSessionHistory(start.SessionHistory)
 		if err != nil {
 			return nil, fmt.Errorf("unable to read JSON from flow start history: %w", err)
 		}
@@ -175,8 +164,8 @@ func StartFlowBatch(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAsse
 
 	// this will build our trigger for each contact started
 	triggerBuilder := func(contact *flows.Contact) flows.Trigger {
-		if !batch.ParentSummary.IsNull() {
-			tb := triggers.NewBuilder(oa.Env(), flow.Reference(), contact).FlowAction(history, json.RawMessage(batch.ParentSummary))
+		if !start.ParentSummary.IsNull() {
+			tb := triggers.NewBuilder(oa.Env(), flow.Reference(), contact).FlowAction(history, json.RawMessage(start.ParentSummary))
 			if batchStart {
 				tb = tb.AsBatch()
 			}
@@ -184,40 +173,39 @@ func StartFlowBatch(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAsse
 		}
 
 		tb := triggers.NewBuilder(oa.Env(), flow.Reference(), contact).Manual()
-		if !batch.Params.IsNull() {
+		if !start.Params.IsNull() {
 			tb = tb.WithParams(params)
 		}
 		if batchStart {
 			tb = tb.AsBatch()
 		}
-		return tb.WithUser(flowUser).WithOrigin(startTypeToOrigin[batch.StartType]).Build()
+		return tb.WithUser(flowUser).WithOrigin(startTypeToOrigin[start.StartType]).Build()
 	}
 
-	// before committing our runs we want to set the start they are associated with
-	updateStartID := func(ctx context.Context, tx *sqlx.Tx, rp *redis.Pool, oa *models.OrgAssets, sessions []*models.Session) error {
-		// for each run in our sessions, set the start id
-		for _, s := range sessions {
-			for _, r := range s.Runs() {
-				r.SetStartID(batch.StartID)
+	// if we have a persisted start, we need to set its id on runs before they are saved
+	var commitHook models.SessionCommitHook
+	if start.ID != models.NilStartID {
+		commitHook = func(ctx context.Context, tx *sqlx.Tx, rp *redis.Pool, oa *models.OrgAssets, sessions []*models.Session) error {
+			// for each run in our sessions, set the start id
+			for _, s := range sessions {
+				for _, r := range s.Runs() {
+					r.SetStartID(batch.StartID)
+				}
 			}
+			return nil
 		}
-		return nil
 	}
 
 	options := &StartOptions{
 		Interrupt:      flow.FlowType().Interrupts(),
 		TriggerBuilder: triggerBuilder,
-		CommitHook:     updateStartID,
+		CommitHook:     commitHook,
 	}
 
 	sessions, err := StartFlow(ctx, rt, oa, flow, batch.ContactIDs, options)
 	if err != nil {
 		return nil, fmt.Errorf("error starting flow batch: %w", err)
 	}
-
-	// log both our total and average
-	analytics.Gauge("mr.flow_batch_start_elapsed", float64(time.Since(start))/float64(time.Second))
-	analytics.Gauge("mr.flow_batch_start_count", float64(len(sessions)))
 
 	return sessions, nil
 }
@@ -264,7 +252,7 @@ func tryToStartWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.Org
 	if err != nil {
 		return nil, nil, err
 	}
-	locked := maps.Keys(locks)
+	locked := slices.Collect(maps.Keys(locks))
 
 	// whatever happens, we need to unlock the contacts
 	defer models.UnlockContacts(rt, oa.OrgID(), locks)
@@ -409,7 +397,7 @@ func StartFlowForContacts(
 			err = tx.Commit()
 			if err != nil {
 				tx.Rollback()
-				log.Error("error comitting session to db", "error", err, "contact_uuid", session.Contact().UUID())
+				log.Error("error committing session to db", "error", err, "contact_uuid", session.Contact().UUID())
 				continue
 			}