+++ core/tasks/task.go
@@ -13,8 +13,9 @@ import (
 	"github.com/nyaruka/mailroom/utils/queues"
 )
 
-var HandlerQueue = queues.NewFairSorted("handler")
-var BatchQueue = queues.NewFairSorted("batch")
+var HandlerQueue = queues.NewFairSorted("tasks:handler")
+var BatchQueue = queues.NewFairSorted("tasks:batch")
+var ThrottledQueue = queues.NewFairSorted("tasks:throttled")
 
 var registeredTypes = map[string](func() Task){}