+++ runtime/config.go
@@ -61,10 +61,12 @@ type Config struct {
 	AWSSecretAccessKey string `help:"secret access key to use for AWS services"`
 	AWSRegion          string `help:"region to use for AWS services, e.g. us-east-1"`
 
+	DynamoEndpoint    string `help:"DynamoDB service endpoint, e.g. https://dynamodb.us-east-1.amazonaws.com"`
+	DynamoTablePrefix string `help:"prefix to use for DynamoDB tables"`
+
 	S3Endpoint          string `help:"S3 service endpoint, e.g. https://s3.amazonaws.com"`
 	S3AttachmentsBucket string `help:"S3 bucket to write attachments to"`
 	S3SessionsBucket    string `help:"S3 bucket to write flow sessions to"`
-	S3LogsBucket        string `help:"S3 bucket to write channel logs to"`
 	S3Minio             bool   `help:"S3 is actually Minio or other compatible service"`
 
 	CourierAuthToken string `help:"the authentication token used for requests to Courier"`
@@ -119,10 +121,12 @@ func NewDefaultConfig() *Config {
 		AWSSecretAccessKey: "",
 		AWSRegion:          "us-east-1",
 
+		DynamoEndpoint:    "", // let library generate it
+		DynamoTablePrefix: "Temba",
+
 		S3Endpoint:          "https://s3.amazonaws.com",
 		S3AttachmentsBucket: "temba-attachments",
 		S3SessionsBucket:    "temba-sessions",
-		S3LogsBucket:        "temba-logs",
 
 		InstanceID: hostname,
 		LogLevel:   slog.LevelWarn,