+++ runtime/runtime.go
@@ -8,7 +8,8 @@ import (
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/gomodule/redigo/redis"
 	"github.com/jmoiron/sqlx"
-	"github.com/nyaruka/gocommon/s3x"
+	"github.com/nyaruka/gocommon/aws/dynamo"
+	"github.com/nyaruka/gocommon/aws/s3x"
 )
 
 // Runtime represents the set of services required to run many Mailroom functions. Used as a wrapper for
@@ -17,6 +18,7 @@ type Runtime struct {
 	DB         *sqlx.DB
 	ReadonlyDB *sql.DB
 	RP         *redis.Pool
+	Dynamo     *dynamo.Service
 	S3         *s3x.Service
 	ES         *elasticsearch.TypedClient
 	FCM        FCMClient