+++ testsuite/testsuite.go
@@ -3,20 +3,25 @@ package testsuite
 import (
 	"context"
 	"fmt"
+	"io"
 	"log/slog"
+	"maps"
 	"os"
 	"os/exec"
 	"path"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/gomodule/redigo/redis"
 	"github.com/jmoiron/sqlx"
-	"github.com/nyaruka/gocommon/s3x"
+	"github.com/nyaruka/gocommon/aws/dynamo"
+	"github.com/nyaruka/gocommon/aws/s3x"
+	"github.com/nyaruka/gocommon/jsonx"
 	"github.com/nyaruka/mailroom/core/models"
 	"github.com/nyaruka/mailroom/runtime"
 	"github.com/nyaruka/redisx/assertredis"
 	"github.com/nyaruka/rp-indexer/v9/indexers"
-	"golang.org/x/exp/maps"
 )
 
 var _db *sqlx.DB
@@ -36,6 +41,7 @@ const (
 	ResetRedis   = ResetFlag(1 << 3)
 	ResetStorage = ResetFlag(1 << 4)
 	ResetElastic = ResetFlag(1 << 5)
+	ResetDynamo  = ResetFlag(1 << 6)
 )
 
 // Reset clears out both our database and redis DB
@@ -56,6 +62,9 @@ func Reset(what ResetFlag) {
 	if what&ResetElastic > 0 {
 		resetElastic(ctx, rt)
 	}
+	if what&ResetDynamo > 0 {
+		resetDynamo(ctx, rt)
+	}
 
 	models.FlushCache()
 }
@@ -70,8 +79,12 @@ func Runtime() (context.Context, *runtime.Runtime) {
 	cfg.S3Endpoint = "http://localhost:9000"
 	cfg.S3AttachmentsBucket = "test-attachments"
 	cfg.S3SessionsBucket = "test-sessions"
-	cfg.S3LogsBucket = "test-logs"
 	cfg.S3Minio = true
+	cfg.DynamoEndpoint = "http://localhost:6000"
+	cfg.DynamoTablePrefix = "Test"
+
+	dyna, err := dynamo.NewService(cfg.AWSAccessKeyID, cfg.AWSSecretAccessKey, cfg.AWSRegion, cfg.DynamoEndpoint, cfg.DynamoTablePrefix)
+	noError(err)
 
 	s3svc, err := s3x.NewService(cfg.AWSAccessKeyID, cfg.AWSSecretAccessKey, cfg.AWSRegion, cfg.S3Endpoint, cfg.S3Minio)
 	noError(err)
@@ -81,8 +94,9 @@ func Runtime() (context.Context, *runtime.Runtime) {
 		DB:         dbx,
 		ReadonlyDB: dbx.DB,
 		RP:         getRP(),
-		ES:         getES(),
+		Dynamo:     dyna,
 		S3:         s3svc,
+		ES:         getES(),
 		FCM:        &MockFCMClient{ValidTokens: []string{"FCMID3", "FCMID4", "FCMID5"}},
 		Config:     cfg,
 	}
@@ -157,7 +171,7 @@ func resetDB() {
 }
 
 func loadTestDump() {
-	dump, err := os.Open(absPath("./mailroom_test.dump"))
+	dump, err := os.Open(absPath("./testsuite/testfiles/postgres.dump"))
 	must(err)
 	defer dump.Close()
 
@@ -198,7 +212,6 @@ func resetRedis() {
 func resetStorage(ctx context.Context, rt *runtime.Runtime) {
 	rt.S3.EmptyBucket(ctx, rt.Config.S3AttachmentsBucket)
 	rt.S3.EmptyBucket(ctx, rt.Config.S3SessionsBucket)
-	rt.S3.EmptyBucket(ctx, rt.Config.S3LogsBucket)
 }
 
 // clears indexed data in Elastic
@@ -212,13 +225,38 @@ func resetElastic(ctx context.Context, rt *runtime.Runtime) {
 		noError(err)
 
 		// and delete them
-		for _, index := range maps.Keys(ar) {
+		for index := range maps.Keys(ar) {
 			_, err := rt.ES.Indices.Delete(index).Do(ctx)
 			noError(err)
 		}
 	}
 
 	ReindexElastic(ctx)
+}
+
+func resetDynamo(ctx context.Context, rt *runtime.Runtime) {
+	tablesFile, err := os.Open(absPath("./testsuite/testfiles/dynamo.json"))
+	must(err)
+	defer tablesFile.Close()
+
+	tablesJSON, err := io.ReadAll(tablesFile)
+	must(err)
+
+	inputs := []*dynamodb.CreateTableInput{}
+	jsonx.MustUnmarshal(tablesJSON, &inputs)
+
+	for _, input := range inputs {
+		input.TableName = aws.String(rt.Dynamo.TableName(*input.TableName))
+
+		// delete table if it exists
+		if _, err := rt.Dynamo.Client.DescribeTable(ctx, &dynamodb.DescribeTableInput{TableName: input.TableName}); err == nil {
+			_, err := rt.Dynamo.Client.DeleteTable(ctx, &dynamodb.DeleteTableInput{TableName: input.TableName})
+			must(err)
+		}
+
+		_, err := rt.Dynamo.Client.CreateTable(ctx, input)
+		must(err)
+	}
 }
 
 var sqlResetTestData = `
@@ -239,12 +277,14 @@ DELETE FROM triggers_trigger WHERE id >= 30000;
 DELETE FROM channels_channel WHERE id >= 30000;
 DELETE FROM channels_channelcount;
 DELETE FROM channels_channelevent;
+DELETE FROM channels_channellog;
 DELETE FROM msgs_msg;
 DELETE FROM flows_flowrun;
 DELETE FROM flows_flowpathcount;
 DELETE FROM flows_flownodecount;
 DELETE FROM flows_flowrunstatuscount;
 DELETE FROM flows_flowcategorycount;
+DELETE FROM flows_flowstartcount;
 DELETE FROM flows_flowstart_contacts;
 DELETE FROM flows_flowstart_groups;
 DELETE FROM flows_flowstart;
@@ -253,6 +293,7 @@ DELETE FROM flows_flowrevision WHERE flow_id >= 30000;
 DELETE FROM flows_flow WHERE id >= 30000;
 DELETE FROM ivr_call;
 DELETE FROM campaigns_eventfire;
+DELETE FROM msgs_systemlabelcount;
 DELETE FROM msgs_msg_labels;
 DELETE FROM msgs_msg;
 DELETE FROM msgs_broadcast_groups;